+++ R/basefunctions.R
@@ -89,7 +89,7 @@ markSign <- function(SignIn, plabel = c("n.s.", "+", "*", "**", "***")) {
   return(SignOut)
 }
 
-#' Re-format p-values, avoiding rounding to 0
+#' Re-format p-values, avoiding rounding to 0 and adding surprisal if requested
 #'
 #' \code{formatP} simplifies p-values by rounding to the maximum of p or a
 #' predefined level. Optionally < or = can be added, as well as
@@ -101,17 +101,21 @@ markSign <- function(SignIn, plabel = c("n.s.", "+", "*", "**", "***")) {
 #' @param pretext Should = or < be added before p (default=FALSE)?
 #' @param mark Should significance level be added after p (default=FALSE)?
 #' @param german_num change dot (default) to comma?
+#' @param add.surprisal Add surprisal aka Shannon information to p-value (default=FALSE)?
+#' @param sprecision Rounding level for surprisal (default=1).
 #'
 #' @returns vector or matrix (depending on type of pIn) with type character (default) or numeric,
 #' depending on parameter textout
 #'
 #' @examples
 #' formatP(0.012345)
+#' formatP(0.012345, surprisal = TRUE)
 #' formatP(0.012345, ndigits = 4)
 #' formatP(0.000122345, ndigits = 3, pretext = TRUE)
 #' @export
 formatP <- function(pIn, ndigits = 3, textout = TRUE, pretext = FALSE, 
-                    mark = FALSE, german_num = FALSE) {
+                    mark = FALSE, german_num = FALSE,
+                    add.surprisal = FALSE, sprecision = 1) {
   decimal.mark <- ifelse(german_num, ",", ".")
   pIn_is_matrix <- is.matrix(pIn)
  if(pIn_is_matrix){
@@ -119,7 +123,7 @@ formatP <- function(pIn, ndigits = 3, textout = TRUE, pretext = FALSE,
  } else{
    pIn <- as.numeric(pIn)
  }
-    formatp <- NA_character_
+  formatp <- NA_character_
   if (length(na.omit(pIn))>0) {
     if (!pIn_is_matrix) {
       pIn <- matrix(pIn)
@@ -156,7 +160,14 @@ formatP <- function(pIn, ndigits = 3, textout = TRUE, pretext = FALSE,
         ncol = ncol(pIn)
       )
     }
-    if (textout == FALSE & pretext == FALSE) {
+    if(add.surprisal){
+      s <- apply(pIn,MARGIN = c(1,2),surprisal, precision = sprecision)
+      if(german_num){
+        s <- str_replace(s,'\\.',',')
+      }
+      formatp <- paste0(formatp,', s = ',s)
+    }
+    if (textout == FALSE & pretext == FALSE & add.surprisal == FALSE) {
       formatp <- apply(formatp, MARGIN = c(1, 2), as.numeric)
     }
     if(!pIn_is_matrix){
@@ -399,4 +410,19 @@ tab.search <- function(searchdata = rawdata, pattern,
     positions <- names(positions)
   }
   return(positions)
+}
+
+#' Compute surprisal aka Shannon information from p-values
+#' 
+#' \code{surprisal} takes p-values and returns s, a value representing the
+#' number of consecutive heads on a fair coin, that would be as surprising 
+#' as the p-value
+#' 
+#' @param p a vector of p-values
+#' @param precision rounding level with default 1
+#'
+#' @return a character vector of s-values
+#' @export
+surprisal <- function(p, precision = 1){
+round(-log2(as.numeric(p)),precision) %>% as.character()
 }